backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.4.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
